refactor(python): Move `PythonInspector` related extension function
Move funtions that build the bridge between PythonInspector's and ORT's
models to a dedicated file in order for `PythonInspector` to be
independent of ORT's model for a cleaner separation.
